+++ packages/text-selection-toolbar/web/src/index.ts
@@ -1,3 +1,5 @@
-import TextSelectionListener from './TextSelectionListener.jsx';
+import TextSelectionToolbar from './TextSelectionToolbar.jsx';
 import ViewerInlineToolBar from './ViewerInlineToolBar.jsx';
-export { TextSelectionListener, ViewerInlineToolBar };
+import TwitterButton from './TwitterButton.jsx';
+
+export { TextSelectionToolbar, ViewerInlineToolBar, TwitterButton };